25 update library to 20250902 of cda and other improvements

+++ scripts/tests2exampledocs.js
@@ -18,6 +18,9 @@ const testDirectory = "./tests/endpoints";
 const genDirectory = "./tests/generator";
 const outputDirectory = "./cwmsjs/docs/examples";
 const templatePath = "./scripts/exampletemplate.html"; // Path to your HTML template
+const packageVersion = JSON.parse(
+  fs.readFileSync("./cwmsjs/package.json", "utf8")
+).version;
 
 // Ensure the output directory exists
 if (!fs.existsSync(outputDirectory)) {
@@ -29,10 +32,11 @@ function getFiles(directories) {
   directories.forEach((d) => {
     all_files = fs
       .readdirSync(d)
+      .filter((fileName) => fileName.endsWith(".test.js"))
       .map((fileName) => path.join(d, fileName))
       .concat(all_files);
   });
-  return all_files;
+  return all_files.sort();
 }
 // Read the HTML template
 fs.readFile(templatePath, "utf8", (err, template) => {
@@ -82,6 +86,7 @@ fs.readFile(templatePath, "utf8", (err, template) => {
           // Replace placeholder in template
           const filledTemplate = template
             .replaceAll("${docName}", docFullName)
+            .replaceAll("${packageVersion}", packageVersion)
             .replaceAll(
               "${pageBody}",
               `
@@ -148,6 +153,7 @@ ${formattedBlock.replaceAll("new ", "new cwmsjs.")}\n</script>`) +
     .replaceAll("Example:", "")
     // .replaceAll(":", "")
     .replaceAll("${docName}", "Home")
+    .replaceAll("${packageVersion}", packageVersion)
     .replaceAll("    ", "")
     .replaceAll(
       "${pageBody}",
@@ -163,36 +169,40 @@ ${formattedBlock.replaceAll("new ", "new cwmsjs.")}\n</script>`) +
   // Update the modules file and index file to have links to the examples
   const indexPath = "./cwmsjs/docs/index.html";
   const modulesPath = "./cwmsjs/docs/modules.html";
-  fs.readFile(modulesPath, "utf8", (err, content) => {
-    if (err) throw err;
-    if (content.indexOf("Examples") >= 0) return;
-    const updatedContent = content.replace(
-      "</h3>",
-      `</h3>
+  if (fs.existsSync(modulesPath)) {
+    fs.readFile(modulesPath, "utf8", (err, content) => {
+      if (err) throw err;
+      if (content.indexOf("Examples") >= 0) return;
+      const updatedContent = content.replace(
+        "</h3>",
+        `</h3>
              <h3 class="tsd-index-heading"><a href="/cwms-data-api-client-javascript/examples/">Examples Home</a></h3>
              ${exampleLinks}
             `
-    );
-    fs.writeFile(modulesPath, updatedContent, (err) => {
-      if (err) throw err;
-      console.log(`Updated modules file: ${modulesPath}`);
+      );
+      fs.writeFile(modulesPath, updatedContent, (err) => {
+        if (err) throw err;
+        console.log(`Updated modules file: ${modulesPath}`);
+      });
     });
-  });
-  fs.readFile(indexPath, "utf8", (err, content) => {
-    if (err) throw err;
-    if (content.indexOf("Examples") >= 0) return;
-    const updatedContent = content.replace(
-      "</h2>",
-      `</h2>
+  }
+  if (fs.existsSync(indexPath)) {
+    fs.readFile(indexPath, "utf8", (err, content) => {
+      if (err) throw err;
+      if (content.indexOf("Examples") >= 0) return;
+      const updatedContent = content.replace(
+        "</h2>",
+        `</h2>
              <p><a href="/cwms-data-api-client-javascript/examples/">Examples Home</a></p>
              ${exampleLinks}
             `
-    );
-    fs.writeFile(indexPath, updatedContent, (err) => {
-      if (err) throw err;
-      console.log(`Updated modules file: ${indexPath}`);
+      );
+      fs.writeFile(indexPath, updatedContent, (err) => {
+        if (err) throw err;
+        console.log(`Updated modules file: ${indexPath}`);
+      });
     });
-  });
+  }
 });
 
 function escapeHtml(unsafe) {
